Added a skip if not equal or larger
The Algorithm is meant to keep queue size up to highest dimension,
which is constant.

Without the skip the algorithm will still be O(n) because it visits
every element only once and do a constant number of things for each
visit. Although the skip will reduce the constant amortised. The
worst-case is still the same.